pyproject.toml
[tool.poetry]
name = "book"
version = "0.1.0"
description = "Desenvolvendo um blog com django."
authors = ["andreFSilva <andrealekhine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
python-decouple = "^3.8"
django-extensions = "^3.2.3"
pillow = "^10.2.0"
psycopg-binary = "^3.1.17"
ignr = "^2.2"
django-taggit = "^5.0.1"
markdown = "^3.5.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
